Moved quorem
